refactor: consolidate duplicate logout handling into useLogout hook
LogoutButton, CompanyProfileSectionContainer, and ProfileSectionContainer
each reimplemented the same confirm+fetch+clear-session logic with no
error handling if the logout request failed. Extract to one hook that
surfaces failures instead of silently doing nothing.
